requirements.txt
streamlit==1.31.0
openai>=1.3.0
python-dotenv==1.0.0
httpx>=0.24.0